File: Add a finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_expenses = int(input("Enter your total monthly expenses: "))
-monthly_income = int(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
 
 monthly_savings = monthly_expenses - monthly_income
 projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
